Update from original master repo

+++ build.sh
@@ -1,16 +1,31 @@
 #!/bin/bash
 
+set -e
+
 # Default flags.
-KSP_VERSION_DEFAULT="1.0.2"
+KSP_VERSION_DEFAULT="1.0.5"
 KSP_NAME_DEFAULT="dummy"
 
 # Locations of CKAN and NetKAN.
 LATEST_CKAN_URL="http://ckan-travis.s3.amazonaws.com/ckan.exe"
 LATEST_NETKAN_URL="http://ckan-travis.s3.amazonaws.com/netkan.exe"
+LATEST_CKAN_META="https://github.com/KSP-CKAN/CKAN-meta/archive/master.tar.gz"
 
 # Third party utilities.
 JQ_PATH="jq"
 
+# Return codes.
+EXIT_OK=0
+EXIT_FAILED_PROVE_STEP=1
+EXIT_FAILED_JSON_VALIDATION=2
+
+# Allow us to specify a commit id as the first argument
+if [ -n "$1" ]
+then
+    echo "Using CLI argument of $1"
+    ghprbActualCommit=$1
+fi
+    
 # ------------------------------------------------
 # Function for creating dummy KSP directories to
 # test on. Takes version as an argument.
@@ -27,15 +42,52 @@ create_dummy_ksp () {
     fi
     
     # TODO: Manual hack, a better way to handle this kind of identifiers may be needed.
-    if [ "$KSP_VERSION" == "0.90" ]
-    then
+    case $KSP_VERSION in
+    "0.23")
+        echo "Overidding '0.23' with '0.23.0'"
+        KSP_VERSION="0.23.0"
+        ;;
+    "0.25")
+        echo "Overidding '0.25' with '0.25.0'"
+        KSP_VERSION="0.25.0"
+        ;;
+    "0.90")
+        echo "Overiding '0.90' with '0.90.0'"
         KSP_VERSION="0.90.0"
-    fi
+        ;;
+    "1.0")
+        echo "Overidding '1.0' with '$KSP_VERSION_DEFAULT'"
+        KSP_VERSION=$KSP_VERSION_DEFAULT
+        ;;
+    "1.0.99")
+        echo "Overidding '1.0.99' with '$KSP_VERSION_DEFAULT'"
+        KSP_VERSION=$KSP_VERSION_DEFAULT
+        ;;
+    "any")
+        echo "Overridding any with '$KSP_VERSION_DEFAULT'"
+        KSP_VERSION=$KSP_VERSION_DEFAULT
+        ;;
+    "null")
+        echo "Overriding 'null' with '$KSP_VERSION_DEFAULT'"
+        KSP_VERSION=$KSP_VERSION_DEFAULT
+        ;;
+    "")
+        echo "Overriding empty version with '$KSP_VERSION_DEFAULT'"
+        KSP_VERSION=$KSP_VERSION_DEFAULT
+        ;;
+    *)
+        echo "No override, Running with '$KSP_VERSION'"
+        ;;
+    esac
     
-    echo "Creating a dummy KSP $KSP_VERSION install"
+    
+    echo "Creating a dummy KSP '$KSP_VERSION' install"
     
     # Remove any existing KSP dummy install.
-    rm -rf dummy_ksp
+    if [ -d "dummy_ksp/" ]
+    then
+        rm -rf dummy_ksp
+    fi
     
     # Create a new dummy KSP.
     mkdir dummy_ksp
@@ -56,7 +108,7 @@ create_dummy_ksp () {
     # Reset the Mono registry.
     if [ "$USER" = "jenkins" ]
     then
-        REGISTRY_FILE=${HOME}/.mono/registry/CurrentUser/software/ckan/values.xml
+        REGISTRY_FILE=$HOME/.mono/registry/CurrentUser/software/ckan/values.xml
         if [ -r $REGISTRY_FILE ]
         then
             rm -f $REGISTRY_FILE
@@ -64,17 +116,17 @@ create_dummy_ksp () {
     fi
     
     # Register the new dummy install.
-    mono ckan.exe ksp add ${KSP_NAME} "`pwd`/dummy_ksp"
+    mono ckan.exe ksp add $KSP_NAME "`pwd`/dummy_ksp"
     
     # Set the instance to default.
-    mono ckan.exe ksp default ${KSP_NAME}
+    mono ckan.exe ksp default $KSP_NAME
     
     # Point to the local metadata instead of GitHub.
     mono ckan.exe repo add local "file://`pwd`/master.tar.gz"
     mono ckan.exe repo remove default
     
     # Link to the downloads cache.
-    ln -s downloads_cache dummy_ksp/CKAN/downloads
+    ln -s ../../downloads_cache/ dummy_ksp/CKAN/downloads/
 }
 
 # ------------------------------------------------
@@ -82,6 +134,10 @@ create_dummy_ksp () {
 # archive. Assummes metadata.tar.gz to be present.
 # ------------------------------------------------
 inject_metadata () {
+    # TODO: Arrays + Bash Functions aren't fun. This needs
+    # Improvement but appears to work. The variables are
+    # available to the called functions.
+
     # Check input, requires at least 1 argument.
     if [ $# -ne 1 ]
     then
@@ -97,9 +153,12 @@ inject_metadata () {
     tar -xzf metadata.tar.gz
     
     # Copy in the files to inject.
-    for f in $1
+    # TODO: Unsure why this suddenly needs [*] declaration
+    # but it does work
+    for f in ${OTHER_FILES[*]}
     do
-        cp f CKAN-meta-master
+        echo "Injecting: $f"
+        cp $f CKAN-meta-master
     done
     
     # Recompress the archive.
@@ -111,11 +170,58 @@ inject_metadata () {
 # Main entry point.
 # ------------------------------------------------
 
+if [ -n "$ghprbActualCommit" ]
+then
+    echo "Commit hash: $ghprbActualCommit"
+    export COMMIT_CHANGES="`git diff --diff-filter=AM --name-only --stat origin/master...HEAD`"
+else
+    echo "No commit provided, skipping further tests."
+    exit $EXIT_OK
+fi
+
 # Make sure we start from a clean slate.
-rm -rf built
-rm -rf downloads_cache
-rm -f master.tar.gz
-rm -f metadata.tar.gz
+if [ -d "built/" ]
+then
+    rm -rf built
+fi
+
+if [ -d "downloads_cache/" ]
+then
+    rm -rf downloads_cache
+fi
+
+if [ -e "master.tar.gz" ]
+then
+    rm -f master.tar.gz
+fi
+
+if [ -e "metadata.tar.gz" ]
+then
+    rm -f metadata.tar.gz
+fi
+
+# Check JSON.
+echo "Running jsonlint on the changed files"
+echo "If you get an error below you should look for syntax errors in the metadata"
+
+for f in $COMMIT_CHANGES
+do
+    if [ "$f" == "build.sh" ]
+    then
+        echo "Lets try not to validate our build script with jsonlint"
+        continue
+    fi
+
+    echo "Validating $f..."
+    jsonlint -s -v $f
+    
+    if [ $? -ne 0 ]
+    then
+        echo "Failed to validate $f"
+        exit $EXIT_FAILED_JSON_VALIDATION
+    fi
+done
+echo ""
 
 # Run basic tests.
 echo "Running basic sanity tests on metadata."
@@ -124,32 +230,19 @@ echo "If these fail, then fix whatever is causing them first."
 if ! prove
 then
     echo "Prove step failed."
-    exit 1
+    exit $EXIT_FAILED_PROVE_STEP
 fi
 
 # Find the changes to test.
 echo "Finding changes to test..."
 
-if [ -z $ghprbActualCommit ]
-then
-    echo "No commit hash, running all netkan files"
-    export COMMIT_CHANGES=NetKAN/*.netkan
-else
-    echo "Commit hash: $ghprbActualCommit"
-    export COMMIT_CHANGES="`git diff --diff-filter=AM --name-only --stat origin/master NetKAN`"
-fi
-
-if [ "$COMMIT_CHANGES" = "" ]
-then
-    echo "No .netkan changes, skipping further tests."
-    exit 0
-fi
-
-# Check JSON.
-echo "Running jsonlint on the changed files"
-echo "If you get an error below you should look for syntax errors in the metadata"
-
-jsonlint -s -v $COMMIT_CHANGES
+# Print the changes.
+echo "Detected file changes:"
+for f in $COMMIT_CHANGES
+do
+    echo "$f"
+done
+echo ""
 
 # Create folders.
 mkdir built
@@ -158,13 +251,15 @@ mkdir downloads_cache # TODO: Point to cache folder here instead if possible.
 # Fetch latest ckan and netkan executable.
 echo "Fetching latest ckan.exe"
 wget --quiet $LATEST_CKAN_URL -O ckan.exe
+mono ckan.exe version
 
 echo "Fetching latest netkan.exe"
 wget --quiet $LATEST_NETKAN_URL -O netkan.exe
+mono netkan.exe --version
 
 # Fetch the latest metadata.
 echo "Fetching latest metadata"
-wget --quiet https://github.com/KSP-CKAN/CKAN-meta/archive/master.tar.gz -O metadata.tar.gz
+wget --quiet $LATEST_CKAN_META -O metadata.tar.gz
 
 # Determine KSP dummy name.
 if [ -z $ghprbActualCommit ]
@@ -174,42 +269,32 @@ else
     KSP_NAME=$ghprbActualCommit
 fi
 
-mono --debug ckan.exe ksp add ${KSP_NAME} "`pwd`/dummy_ksp"
-mono --debug ckan.exe ksp default ${KSP_NAME}
-
-echo Running ckan update
-mono --debug ckan.exe update
-
-echo Running jsonlint on the changed files
-echo If you get an error below you should look for syntax errors in the metadata
-
-jsonlint -s -v ${COMMIT_CHANGES}
-
-echo Fetching latest netkan.exe
-
-# fetch latest netkan.exe (corresponding to CKAN/master)
-wget --quiet https://ckan-travis.s3.amazonaws.com/netkan.exe
-
-mkdir built
-
 # Build all the passed .netkan files.
 # Note: Additional NETKAN_OPTIONS may be set on jenkins jobs
 for f in $COMMIT_CHANGES
 do
+    if [ "$f" = "build.sh" ]; then
+        echo "Lets try not to validate our build script with netkan"
+        continue
+    fi
+
     echo "Running NetKAN for $f"
-    mono netkan.exe $f --cachedir="downloads_cache" --outputdir="built" ${NETKAN_OPTIONS}
+    mono netkan.exe $f --cachedir="downloads_cache" --outputdir="built" $NETKAN_OPTIONS
 done
 
-# Reset KSP.
-mono ckan.exe ksp forget $KSP_NAME_DEFAULT
-
 # Test all the built files.
-for f in built/*.ckan
+for ckan in built/*.ckan
 do
-    echo "Checking $f"
+    if [ ! -e "$ckan" ]
+    then
+        echo "No ckan files to test"
+        continue
+    fi
+
+    echo "Checking $ckan"
     echo "----------------------------------------------"
     echo ""
-    cat $f | python -m json.tool
+    cat $ckan | python -m json.tool
     echo "----------------------------------------------"
     echo ""
     
@@ -218,18 +303,15 @@ do
     
     for o in built/*.ckan
     do
-        if [ "$f" != "$o" ]
-        then
-            OTHER_FILES+=($o)
-        fi
+        OTHER_FILES+=($o)
     done
     
     # Inject into metadata.
     inject_metadata $OTHER_FILES
     
     # Extract identifier and KSP version.
-    CURRENT_IDENTIFIER=$($JQ_PATH '.identifier' $f)
-    CURRENT_KSP_VERSION=$($JQ_PATH 'if .ksp_version then .ksp_version else .ksp_version_min end' $f)
+    CURRENT_IDENTIFIER=$($JQ_PATH '.identifier' $ckan)
+    CURRENT_KSP_VERSION=$($JQ_PATH 'if .ksp_version then .ksp_version else .ksp_version_max end' $ckan)
     
     # Strip "'s.
     CURRENT_IDENTIFIER=${CURRENT_IDENTIFIER//'"'}
@@ -244,8 +326,8 @@ do
     echo "Running ckan update"
     mono ckan.exe update
     
-    echo "Running ckan install -c $f"
-    mono ckan.exe install -c $f --headless
+    echo "Running ckan install -c $ckan"
+    mono ckan.exe install -c $ckan --headless
     
     # Print list of installed mods.
     mono ckan.exe list --porcelain
